--- a/config/app.config
+++ b/config/app.config
@@ -12,7 +12,6 @@ work.directory = ""
 stats.log.dir = '${sys:catalina.base}/logs/statistics'
 alerts.dir = ""
 
-debug.mode = false
 ignore.auth=true
 
 /*  keep vis.shared.mem.size small if pct is used */