--- a/src/firefly/java/edu/caltech/ipac/table/IpacTableUtil.java
+++ b/src/firefly/java/edu/caltech/ipac/table/IpacTableUtil.java
@@ -6,7 +6,6 @@ package edu.caltech.ipac.table;
 import edu.caltech.ipac.util.CollectionUtil;
 import edu.caltech.ipac.util.FileUtil;
 import edu.caltech.ipac.util.StringUtils;
-import org.json.simple.JSONArray;
 import org.json.simple.JSONValue;
 
 import java.io.BufferedReader;
@@ -267,13 +266,16 @@ public class IpacTableUtil {
         if (line != null && line.startsWith("|")) {
             String[] names = parseHeadings(line.trim());
             int cursor = 1;
-            String cname;
+            String cname; String cval;
             for (int idx = 0; idx < names.length; idx++) {
-                cname = names[idx];
-                DataType dt = new DataType(cname.trim(), null);
+                cval = names[idx];
+                cname = cval.trim();
+                DataType dt = new DataType(cname, null);
                 cols.add(dt);
-                tableDef.setColOffsets(idx, cursor);
-                cursor += cname.length() + 1;
+                TableUtil.ParsedColInfo pci = tableDef.getParsedInfo(cname);
+                pci.startIdx = cursor;
+                pci.endIdx = cursor + cval.length();
+                cursor = pci.endIdx + 1;
             }
         }
         tableDef.setCols(cols);
@@ -322,13 +324,13 @@ public class IpacTableUtil {
         }
     }
 
-    public static void applyGuessLogic(DataType type, String val, TableUtil.CheckInfo chkInfo) {
+    public static void applyGuessLogic(DataType type, String val, TableUtil.ParsedColInfo pci) {
 
-        if (!chkInfo.formatChecked) {
-            chkInfo.formatChecked = guessFormatInfo(type, val);
+        if (!pci.formatChecked) {
+            pci.formatChecked = guessFormatInfo(type, val);
         }
 
-        if (!chkInfo.htmlChecked) {
+        if (!pci.htmlChecked) {
             // disable sorting if value is HTML, or unit is 'html'
             // this block should only be executed once, when formatInfo is not set.
             if (type.getDataType() == String.class ) {
@@ -338,7 +340,7 @@ public class IpacTableUtil {
                     type.setFilterable(false);
                 }
             }
-            chkInfo.htmlChecked = true;
+            pci.htmlChecked = true;
         }
     }
 
@@ -400,7 +402,7 @@ public class IpacTableUtil {
     public static DataObject parseRow(DataGroup source, String line, IpacTableDef tableDef) {
         if (line==null) return null;
         DataType[] headers = source.getDataDefinitions();
-        int offset=0, endOfLine=0, endoffset=0;
+        int endOfLine=0, endoffset=0;
         try {
             if (line.startsWith(" ") && line.trim().length() > 0) {
                 DataObject row = new DataObject(source);
@@ -409,40 +411,28 @@ public class IpacTableUtil {
                 DataType dt;
                 for (int idx = 0; idx < headers.length; idx++) {
                     dt = headers[idx];
-                    offset = tableDef.getColOffset(idx);
-                    if (offset > endOfLine) {
+                    TableUtil.ParsedColInfo pci = tableDef.getParsedInfo(dt.getKeyName());
+                    if (pci.startIdx > endOfLine) {
                         // it's okay.  we'll take what's given and treat the rest as null.
                         break;
                     }
-                    endoffset = idx < headers.length-1 ? tableDef.getColOffset(idx+1) : endOfLine;
-
                     // if ending spaces are missing... just ignore it.
-                    if (endoffset > endOfLine) {
-                        endoffset = endOfLine;
-                    }
+                    endoffset = Math.min(pci.endIdx, endOfLine);
 
-                    String val = line.substring(offset, endoffset).trim();
+                    String val = line.substring(pci.startIdx, endoffset).trim();
                     if (!dt.isKnownType()) {
                         IpacTableUtil.guessDataType(dt, val);
                     }
-
-                    TableUtil.CheckInfo checkInfo = tableDef.getColCheckInfos().getCheckInfo(dt.getKeyName());
-                    if (!String.valueOf(tableDef.getAttribute("fixlen")).trim().equals("T")) {
-                        checkInfo.formatChecked = true;     // if fixlen != T, don't guess format
-                    }
-                    applyGuessLogic(dt, val, checkInfo);
+                    applyGuessLogic(dt, val, pci);
 
                     row.setDataElement(dt, dt.convertStringToData(val));
-
-                    offset = endoffset;
                 }
                 return row;
             } else if (line.trim().length() > 0 && !line.startsWith("\\") && !line.startsWith("|")) {
                 throw new RuntimeException("Data row must start with a space.");
             }
         } catch (StringIndexOutOfBoundsException e) {
-            throw new StringIndexOutOfBoundsException("offset="+offset+",endoffset="+endoffset+
-                    ",line.length()="+line.length()+",line="+line);
+            throw new StringIndexOutOfBoundsException("line.length()="+line.length()+",line="+line);
         }
         return null;
     }
@@ -468,6 +458,13 @@ public class IpacTableUtil {
         } finally {
             FileUtil.silentClose(reader);
         }
+    }
+    public static IpacTableDef getMetaInfo(File inf, Map<String, String> metaInfo) throws IOException {
+        IpacTableDef tableDef = getMetaInfo(inf);
+        if (metaInfo != null) {
+            metaInfo.entrySet().forEach((e -> tableDef.setAttribute(e.getKey(), e.getValue())));
+        }
+        return tableDef;
     }
 
     public static IpacTableDef getMetaInfo(BufferedReader reader) throws IOException {