--- a/src/firefly/java/edu/caltech/ipac/util/cache/CacheManager.java
+++ b/src/firefly/java/edu/caltech/ipac/util/cache/CacheManager.java
@@ -55,6 +55,7 @@ public class CacheManager {
     }
 
     public static boolean setCacheProvider(String cacheProviderClassName) {
+        if (isDisabled) return true;
         setCacheProvider(newInstanceOf(cacheProviderClassName));
         return cacheProvider != null;
     }