ticket: https://jira.ipac.caltech.edu/browse/FIREFLY-978
test: https://fireflydev.ipac.caltech.edu/firefly-978-big-table-performance/firefly/

Improve reading IPAC table
- Replace the use of DataGroup and DataObject with indexed values when reading IPAC table to DataGroup
- Greatly reduces read time for large table. (~30%)
- The sample file used to take over 7s now takes around 4.5s to read.

Improve loading IPAC table info DB
- turn off DB logging.  reduce DB ingestion time from 6s to 3s

Lower memory requirement
A sample 1 million rows WISE catalog is around 360mb file; requiring 550mb of RAM in memory.
To ingest the table into the DB requires 3 copies in memory; DataGroup, PrepareStatement, and the DB tables in memory.
  - Instead of loading the full table in one batch commit, separate the table into multiple smaller batches(set to 10000)
  - A PrepareStatement of 10k rows take less memory than having the 3rd copy
  - This slightly decreases loading performance, but uses less memory.
   
Also:
- expand Test functionalities
  - added util functions
  - enhance logging capability
- remove and clean unused flag ‘debug.mode’