@robyww 2022-04-08
> This slightly decreases loading performance, but uses less memory.

It this our goal? Would should try to keep long term memory load  down but the app should be sized for transient memory during a load.