@gpdf 2022-04-08
For Rubin use and for future IRSA use we are going to need to make sure that we are also performant on reading in large VOTables.  Million-row query results are to be expected.  Firefly already performs reasonably well on one-million-row tables, once loaded, so that should be sort of a benchmark to aim at for improving the performance of loads.