@loitly 2022-04-08
> > This slightly decreases loading performance, but uses less memory.
> 
> Is this our goal? Would should try to keep long term memory load down but the app should be sized for transient memory during a load.

I've played around with batchSize on this 1m rows sample file to find what's best and settled at 10k.
Below is the best result from 3 runs for each case.  As you can see, `ingestDataIntoDb` time is pretty much identical and yet the 10k size requires 300mb less to load.  Also, this is a single run.  In the multi users runtime environment, execution time is less important then unpredictable hard limit where the application would error on out-of-memory.

`10k batchSize` 
```
09:18:38.348 [main] INFO  edu.caltech.ipac.util.download.URLDownload - Download Complete: IpacTableFromSource5687858460056933396.tbl : 375256168 bytes, time: 252 ms
    file:/Users/loi/dev/cm/firefly_test_data/large-ipac-tables/wise-950000.tbl
09:18:42.576 [main] DEBUG edu.caltech.ipac.firefly.server.util.StopWatch - fetchDataGroup: IpacTableFromSource ran with elapsed time of 4.4870 SECONDS
09:18:45.525 [main] DEBUG edu.caltech.ipac.firefly.server.util.StopWatch - ingestDataIntoDb: IpacTableFromSource ran with elapsed time of 2.9350 SECONDS
09:18:45.527 [main] DEBUG edu.caltech.ipac.firefly.server.util.StopWatch - createDbFile: IpacTableFromSource ran with elapsed time of 7.7500 SECONDS
09:18:45.541 [main] TRACE edu.caltech.ipac.firefly.server.db.EmbeddedDbUtil - execQuery => SQL: select * from DATA   limit 100 offset 0
09:18:45.585 [main] TRACE edu.caltech.ipac.firefly.server.db.EmbeddedDbUtil - converting a 100 rows ResultSet into a DataGroup
09:18:45.595 [main] DEBUG edu.caltech.ipac.firefly.server.util.StopWatch - getDataset: IpacTableFromSource ran with elapsed time of 0.0670 SECONDS
```

`single batch (1m rows)`
```
09:24:20.381 [main] INFO  edu.caltech.ipac.util.download.URLDownload - Download Complete: IpacTableFromSource830938347332589139.tbl : 375256168 bytes, time: 232 ms
    file:/Users/loi/dev/cm/firefly_test_data/large-ipac-tables/wise-950000.tbl
09:24:24.887 [main] DEBUG edu.caltech.ipac.firefly.server.util.StopWatch - fetchDataGroup: IpacTableFromSource ran with elapsed time of 4.7460 SECONDS
09:24:27.885 [main] DEBUG edu.caltech.ipac.firefly.server.util.StopWatch - ingestDataIntoDb: IpacTableFromSource ran with elapsed time of 2.9830 SECONDS
09:24:27.887 [main] DEBUG edu.caltech.ipac.firefly.server.util.StopWatch - createDbFile: IpacTableFromSource ran with elapsed time of 8.0600 SECONDS
09:24:27.902 [main] TRACE edu.caltech.ipac.firefly.server.db.EmbeddedDbUtil - execQuery => SQL: select * from DATA   limit 100 offset 0
09:24:27.948 [main] TRACE edu.caltech.ipac.firefly.server.db.EmbeddedDbUtil - converting a 100 rows ResultSet into a DataGroup
09:24:27.958 [main] DEBUG edu.caltech.ipac.firefly.server.util.StopWatch - getDataset: IpacTableFromSource ran with elapsed time of 0.0710 SECONDS

```