@loitly 2022-04-08
@ejoliet 950k rows sample file.  firefly_test_data/large-ipac-tables/wise-950000.tbl