@loitly 2022-04-08
@gpdf I agree on VOTable.  In this round, I looked at reading IPAC table, mostly because I know there's room for improvement.  But, I also looked at loading the in-memory object(DataGroup) and ingesting it into the Database which is common to all formats.

In a later round, I will look into reading FITS table, ascii VOTable, and binary VOTable.  But, those relies on 3rd party libraries.  I am not sure how much we can improve without drastic changes.  It's still worth looking into, though.